Fix a bug for a particulary wpml setting
Continue the loop if the variable $otherID is null on the merge_comments function. 
This can happen if the post type translation is set to 'translatable - use translation if available or fallback to default language' on the wpml settings page.
In this case, the get_comments function returns all comments from all posts.